// src/app/models/report.ts
export class Report {
    nome_ganhador: string;
    cpf_ganhador: string;
    telefone_ganhador: string;
    numero_da_sorte: string;
    associado: string;
    telefone_associado: string;
    loja: string;
    cpnj: string;
    endereco: string;
    bairro: string;
    municipio: string;
    id_sorteio: string;
    data_sorteio: string;
  }
  